more progress towards separated nock playback record 7
